FCM token 저장하는 것을 현재 디바이스가 android일 때만 실행하도록 변경하여 아이폰에서도 접속 가능하게 함

+++ back-end/chatting_app/src/main/java/com/giho/chatting_app/service/FriendService.java
@@ -157,14 +157,15 @@ public class FriendService {
       // 상대방이 만든 ChatRoom
       chatRoomOpt = chatRoomRepository.findByCreatorIdAndVisitorId(friendId, myId);
     }
-    ChatRoom chatRoom = chatRoomOpt.orElseThrow(() -> new CustomException(ErrorCode.CHATROOM_NOT_FOUND));
-    
-    // 채팅방에 있는 메세지 삭제
-    List<ChatMessages> chatMessages = chatMessagesRepository.findByRoomId(chatRoom.getId());
-    chatMessagesRepository.deleteAll(chatMessages);
 
-    // 채팅방 삭제
-    chatRoomRepository.delete(chatRoom);
+    if (chatRoomOpt.isPresent()) {
+      // 채팅방에 있는 메세지 삭제
+      List<ChatMessages> chatMessages = chatMessagesRepository.findByRoomId(chatRoomOpt.get().getId());
+      chatMessagesRepository.deleteAll(chatMessages);
+
+      // 채팅방 삭제
+      chatRoomRepository.delete(chatRoomOpt.get());
+    }
 
     // 친구 삭제
     friendRepository.delete(friend);
